move privacy settings button setup function to main.js
