+++ assets/javascripts/modules/analytics/cmp.js
@@ -1,7 +1,4 @@
-import { onConsentChange, cmp } from '@guardian/consent-management-platform';
-
-const PRIVACY_SETTINGS_SELECTOR = '.js-privacy-settings-link';
-const HIDDEN_CLASS = 'is-hidden';
+import { onConsentChange } from '@guardian/consent-management-platform';
 
 const getConsentForVendors = (cmpVendorIds) => new Promise((resolve) => {
     if (!Array.isArray(cmpVendorIds)) {
@@ -37,12 +34,4 @@ const checkCCPA = () => new Promise((resolve) => {
 
 const registerCallbackOnConsentChange = (fn) => onConsentChange(fn);
 
-const createPrivacySettingsLink = () => {
-    const privacySettingsButton = document.querySelector(PRIVACY_SETTINGS_SELECTOR);
-    if (privacySettingsButton) {
-        privacySettingsButton.classList.remove(HIDDEN_CLASS);
-        privacySettingsButton.addEventListener('click', cmp.showPrivacyManager);
-    }
-}
-
-export { getConsentForVendors, checkAllTCFv2PurposesAreOptedIn, checkCCPA, registerCallbackOnConsentChange, createPrivacySettingsLink };
+export { getConsentForVendors, checkAllTCFv2PurposesAreOptedIn, checkCCPA, registerCallbackOnConsentChange };
